requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
black==24.10.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.7
cron-descriptor==1.4.5
Django==5.1.3
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-stubs==5.1.1
django-stubs-ext==5.1.1
django-timezone-field==7.0
flake8==7.1.1
idna==3.10
isort==5.13.2
kombu==5.4.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytils==0.4.1
pytz==2024.2
redis==5.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
